all variants now use the same pld evaluation

+++ tests/test_compute_delta_var.py
@@ -27,7 +27,7 @@ class compute_delta_regression_tests(unittest.TestCase):
         nc = 10
         sigmas = np.linspace(1.6, 1.8, nc)
         q_values = np.linspace(0.05, 0.06, nc)
-        ks = 10 * np.ones(nc) #number of compositions for each value of (q,sigma)
+        ks = 10 * np.ones(nc, dtype=np.int32) #number of compositions for each value of (q,sigma)
 
         test_data = [
             (dict(sigma_t=sigmas, q_t=q_values, k=ks, target_eps=1.0, nx=1E6, L=20.0), 0.0007824006722393885),
@@ -43,16 +43,16 @@ class compute_delta_regression_tests(unittest.TestCase):
 
     def test_get_delta_R_invalid_sizes(self):
         with self.assertRaises(ValueError):
-            get_delta_R(sigma_t = np.ones(2), q_t = np.ones(2), k = np.ones(1))
-            get_delta_R(sigma_t = np.ones(2), q_t = np.ones(1), k = np.ones(2))
-            get_delta_R(sigma_t = np.ones(1), q_t = np.ones(2), k = np.ones(2))
+            get_delta_R(sigma_t = np.ones(2), q_t = np.ones(2), k = np.ones(1, dtype=np.int32))
+            get_delta_R(sigma_t = np.ones(2), q_t = np.ones(1), k = np.ones(2, dtype=np.int32))
+            get_delta_R(sigma_t = np.ones(1), q_t = np.ones(2), k = np.ones(2, dtype=np.int32))
 
     def test_get_delta_S_regression_valid_params(self):
         """ Tests that results of get_delta_S did not drift from last version."""
         nc = 10
         sigmas = np.linspace(1.6, 1.8, nc)
         q_values = np.linspace(0.05, 0.06, nc)
-        ks = 10 * np.ones(nc) #number of compositions for each value of (q,sigma)
+        ks = 10 * np.ones(nc, dtype=np.int32) #number of compositions for each value of (q,sigma)
 
         test_data = [
             (dict(sigma_t=sigmas, q_t=q_values, k=ks, target_eps=1.0, nx=1E6, L=20.0), 0.026891238748125684),
@@ -67,9 +67,9 @@ class compute_delta_regression_tests(unittest.TestCase):
 
     def test_get_delta_S_invalid_sizes(self):
         with self.assertRaises(ValueError):
-            get_delta_S(sigma_t = np.ones(2), q_t = np.ones(2), k = np.ones(1))
-            get_delta_S(sigma_t = np.ones(2), q_t = np.ones(1), k = np.ones(2))
-            get_delta_S(sigma_t = np.ones(1), q_t = np.ones(2), k = np.ones(2))
+            get_delta_S(sigma_t = np.ones(2), q_t = np.ones(2), k = np.ones(1, dtype=np.int32))
+            get_delta_S(sigma_t = np.ones(2), q_t = np.ones(1), k = np.ones(2, dtype=np.int32))
+            get_delta_S(sigma_t = np.ones(1), q_t = np.ones(2), k = np.ones(2, dtype=np.int32))
 
 
 if __name__ == '__main__':
